separated interaction types into their own event files
