+++ events/buttonInteraction.js
@@ -1,0 +1,5 @@
+module.exports = {
+    async buttonExecute(client, interaction) {
+
+    }
+}
